+++ framework/contracts/account/proxy/Cargo.toml
@@ -35,7 +35,6 @@ semver = { workspace = true }
 serde = { workspace = true }
 thiserror = { workspace = true }
 abstract-macros = { workspace = true }
-abstract-ica = { workspace = true }
 
 [target.'cfg(not(target_arch = "wasm32"))'.dependencies]
 workspace-hack = { version = "0.1", path = "../../../workspace-hack" }
