+++ framework/contracts/account/proxy/src/commands.rs
@@ -95,7 +95,7 @@ pub fn ica_action(deps: DepsMut, msg_info: MessageInfo, action_query: Binary) ->
             ))
         })?;
 
-    let res: abstract_ica::msg::IcaActionResult = deps.querier.query(
+    let res: abstract_std::ica_client::IcaActionResult = deps.querier.query(
         &WasmQuery::Smart {
             contract_addr: ica_client_address.into(),
             msg: action_query,
@@ -521,7 +521,7 @@ mod test {
     }
 
     mod ica_action {
-        use abstract_ica::msg::IcaActionResult;
+        use abstract_std::ica_client::IcaActionResult;
         use abstract_std::{manager, proxy::state::State};
 
         use super::*;
