+++ framework/contracts/native/ica-client/Cargo.toml
@@ -27,10 +27,9 @@ abstract-std = { workspace = true }
 abstract-sdk = { workspace = true }
 semver = { workspace = true }
 polytone = { workspace = true }
-abstract-ica = { workspace = true }
-abstract-macros ={ workspace = true }
+abstract-macros = { workspace = true }
 
-evm-note = { version = "0.0.2", features = ["library"]}
+evm-note = { version = "0.0.2", features = ["library"] }
 polytone-evm = "0.0.2"
 
 [target.'cfg(not(target_arch = "wasm32"))'.dependencies]
@@ -40,6 +39,7 @@ workspace-hack = { version = "0.1", path = "../../../workspace-hack" }
 cosmwasm-schema = { workspace = true }
 abstract-testing = { workspace = true }
 speculoos = { workspace = true }
+union-connector = { git = "https://github.com/abstractsdk/evm-ibc", branch = "development/create-union-bundle" }
 
 [profile.release]
 rpath = false
