--- framework/packages/abstract-ica/README.md
@@ -1,3 +1,0 @@
-# Abstract ICA
-
-This crate contains the types required to interact with Interchain Accounts (ICAs) on different chains, using different technologies. It is used by the Abstract IBC Client to interact with ICAs on chains that support it.
