--- framework/packages/abstract-ica/src/lib.rs
@@ -1,9 +1,0 @@
-mod action;
-mod chain_type;
-pub mod msg;
-
-pub use action::{IcaAction, IcaActionResponse, IcaExecute};
-pub use chain_type::{CastChainType, ChainType};
-
-pub use polytone_evm::EVM_NOTE_ID;
-pub use polytone_evm::POLYTONE_EVM_VERSION;
