+++ framework/packages/abstract-interface/src/native/ica_client.rs
@@ -1,0 +1,51 @@
+use abstract_std::{
+    ibc_client::{ExecuteMsg, InstantiateMsg, MigrateMsg, QueryMsg},
+    ICA_CLIENT,
+};
+
+use cw_orch::{contract::Contract, interface, prelude::*};
+
+use crate::RegisteredModule;
+
+#[interface(InstantiateMsg, ExecuteMsg, QueryMsg, MigrateMsg)]
+pub struct IcaClient<Chain>;
+
+impl<Chain: CwEnv> Uploadable for IcaClient<Chain> {
+    #[cfg(feature = "integration")]
+    fn wrapper() -> <Mock as TxHandler>::ContractSource {
+        Box::new(
+            ContractWrapper::new_with_empty(
+                ::ibc_client::contract::execute,
+                ::ibc_client::contract::instantiate,
+                ::ibc_client::contract::query,
+            )
+            .with_migrate(::ibc_client::contract::migrate),
+        )
+    }
+    fn wasm(_chain: &ChainInfoOwned) -> WasmPath {
+        artifacts_dir_from_workspace!()
+            .find_wasm_path("ibc_client")
+            .unwrap()
+    }
+}
+
+impl<Chain: CwEnv> RegisteredModule for IcaClient<Chain> {
+    type InitMsg = cosmwasm_std::Empty;
+
+    fn module_id<'a>() -> &'a str {
+        ICA_CLIENT
+    }
+    fn module_version<'a>() -> &'a str {
+        ibc_client::contract::CONTRACT_VERSION
+    }
+
+    fn dependencies<'a>() -> &'a [abstract_std::objects::dependency::StaticDependency] {
+        &[]
+    }
+}
+
+impl<Chain: CwEnv> From<Contract<Chain>> for IcaClient<Chain> {
+    fn from(value: Contract<Chain>) -> Self {
+        IcaClient(value)
+    }
+}
