+++ framework/packages/abstract-interface/src/native/mod.rs
@@ -2,6 +2,7 @@ mod account_factory;
 mod ans_host;
 mod ibc_client;
 mod ibc_host;
+mod ica_client;
 mod module_factory;
 mod version_control;
