+++ framework/packages/abstract-std/src/native/mod.rs
@@ -3,5 +3,6 @@ pub mod ans_host;
 pub mod ibc;
 pub mod ibc_client;
 pub mod ibc_host;
+pub mod ica_client;
 pub mod module_factory;
 pub mod version_control;
